Rework requesting and handling experimental/serverStatus
Add a command; remove two unnecessary `advice'.

* eglot-x.el (eglot-x--put-in-server, eglot-x--get-from-server):
Simply store and retrieve the state of eglot-x in the server
object and not try to hide this from Eglot.
(eglot-x--remove-hidden-info)
(eglot-x--hack-mode-line-format): Delete.
(eglot-x-setup, eglot-x-disable): Don't use the deleted defuns.
(eglot-x--hack-mode-line-format): Rename to
eglot-x--mode-line-format and simplify.
(eglot-x-show-server-status): New command.
(eglot-x-menu): Add the command to it.
(eglot-handle-notification:experimental/serverStatus): Update the
mode-line of every window as they might connected to the current
server.